refactor: newsletter to the correct component
